+++ netbox_healthcheck_plugin/views.py
@@ -1,6 +1,10 @@
-from django.views.generic import View
-from health_check.views import MainView
+from health_check.views import HealthCheckView
 
 
-class HealthCheckListView(MainView):
+class HealthCheckListView(HealthCheckView):
     template_name = 'netbox_healthcheck_plugin/healthcheck.html'
+    checks = [
+        'health_check.Database',
+        'netbox_healthcheck_plugin.backends.redis.NetBoxRedisCacheHealthCheck',
+        'netbox_healthcheck_plugin.backends.redis.NetBoxRedisTasksHealthCheck',
+    ]